feat: rewriting the library in typescript
