+++ .eslintrc
@@ -1,8 +1,22 @@
 {
-  "extends": "@readme/eslint-config",
+  "extends": [
+    "@readme/eslint-config",
+    "@readme/eslint-config/docs",
+    "@readme/eslint-config/typescript"
+  ],
+  "plugins": ["jsdoc"],
   "root": true,
   "rules": {
-    "consistent-return": "off",
-    "no-else-return": "off"
+    "@typescript-eslint/no-explicit-any": "off",
+
+    "camelcase": ["error", { "allow": ["OpenAPIV3_1"] }],
+
+    "jsdoc/require-jsdoc": "error",
+    "jsdoc/require-param": "off",
+    "jsdoc/require-param-type": "off",
+    "jsdoc/require-returns": "off",
+    "jsdoc/require-returns-type": "off",
+
+    "no-case-declarations": "off"
   }
 }
