+++ .github/workflows/ci.yml
@@ -19,4 +19,5 @@ jobs:
           node-version: ${{ matrix.node-version }}
 
       - run: npm ci
+      - run: npm run build
       - run: npm test
