+++ README.md
@@ -14,22 +14,23 @@ npm install oas-normalize --save
 
 # Usage
 
-It's pretty simple:
-
 ```javascript
-const OASNormalize = require('oas-normalize');
+import OASNormalize from 'oas-normalize';
 
 const oas = new OASNormalize(
   // Or a string, pathname, JSON blob, whatever
   'https://raw.githubusercontent.com/OAI/OpenAPI-Specification/master/examples/v3.0/petstore-expanded.yaml'
 );
 
-oas.validate().then(definition => {
-  // Definition will always be JSON, and valid.
-  console.log(definition);
-}).catch(err => {
-  console.log(err);
-});
+oas
+  .validate()
+  .then(definition => {
+    // Definition will always be JSON, and valid.
+    console.log(definition);
+  })
+  .catch(err => {
+    console.log(err);
+  });
 ```
 
 # Errors
@@ -70,7 +71,7 @@ If you want some more functionality, you can do anything here:
 
 ### Always Return OpenAPI 3.x
 
-If you want `.validate` to always return an OpenAPI 3.x definition, supply `true` as its argument:
+If you want `.validate()` to always return an OpenAPI 3.x definition, supply `true` as its argument:
 
 ```js
 OASNormalize.validate(true).then(...);
@@ -81,7 +82,7 @@ OASNormalize.validate(true).then(...);
 For security reasons, you need to opt into allowing fetching by a local path. To enable it supply the `enablePaths` option to the class instance:
 
 ```js
-const oas = new OASNormalize('./petstore.json', { enablePaths: true })
+const oas = new OASNormalize('./petstore.json', { enablePaths: true });
 ```
 
 ### Colorized errors
@@ -91,7 +92,7 @@ If you wish errors from `.validate()` to be styled and colorized, supply `colori
 ```js
 const oas = new OASNormalize('https://example.com/petstore.json', {
   colorizeErrors: true,
-})
+});
 ```
 
 Error messages will look like such:
