+++ __tests__/.eslintrc
@@ -1,3 +1,6 @@
 {
-  "extends": "@readme/eslint-config/testing"
+  "extends": "@readme/eslint-config/testing",
+  "rules": {
+    "jsdoc/require-jsdoc": "off"
+  }
 }
