+++ package.json
@@ -2,7 +2,8 @@
   "name": "oas-normalize",
   "version": "6.0.0",
   "description": "OpenAPI 3.x or Swagger 2.0? YAML or JSON? URL, path, string or object? Who cares! It just works.",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "engines": {
     "node": ">=14"
   },
@@ -29,25 +30,35 @@
     "url": "https://github.com/readmeio/oas-normalize/issues"
   },
   "scripts": {
-    "lint": "eslint .",
+    "build": "tsc",
+    "lint": "eslint . --ext .js,.ts",
+    "prebuild": "rm -rf dist/",
+    "prepack": "npm run build",
     "pretest": "npm run lint",
-    "prettier": "prettier --list-different --write \"./**/**.{js,jsx}\"",
-    "test": "jest --coverage"
+    "prettier": "prettier --list-different --write \"./**/**.{js,ts}\"",
+    "test": "tsc; jest --coverage"
   },
   "license": "MIT",
   "dependencies": {
     "@readme/openapi-parser": "^2.2.0",
     "js-yaml": "^4.1.0",
     "node-fetch": "^2.6.1",
+    "openapi-types": "^12.0.0",
     "swagger2openapi": "^7.0.8"
   },
   "devDependencies": {
-    "@readme/eslint-config": "^9.0.0",
+    "@readme/eslint-config": "^10.0.0",
     "@readme/oas-examples": "^5.1.1",
-    "eslint": "^8.13.0",
+    "@types/jest": "^28.1.6",
+    "@types/js-yaml": "^4.0.5",
+    "@types/node-fetch": "^2.6.2",
+    "eslint": "^8.21.0",
     "jest": "^28.0.3",
+    "jet": "^0.8.1",
     "nock": "^13.2.4",
-    "prettier": "^2.6.2"
+    "prettier": "^2.7.1",
+    "ts-jest": "^28.0.7",
+    "typescript": "^4.7.4"
   },
   "prettier": "@readme/eslint-config/prettier"
 }
