+++ src/.sink.d.ts
@@ -1,0 +1,2 @@
+// These packages don't have any TS types so we need to declare a module in order to use them.
+declare module 'swagger2openapi';
