+++ tsconfig.json
@@ -1,0 +1,14 @@
+{
+  "compilerOptions": {
+    "allowJs": true,
+    "baseUrl": "./src",
+    "declaration": true,
+    "esModuleInterop": true,
+    "lib": ["dom", "es2020"],
+    "outDir": "dist/",
+    "paths": {
+      "swagger2openapi": [".sink.d.ts"]
+    },
+  },
+  "include": ["./src/**/*"]
+}
